ads_project/requirement.txt
django==3.2
django-ckeditor
requests
AdvancedHTMLParser
Pillow